Multi-cast skills now listed within unit activated skills list (#496).
